LAVA JA FA MAL

+++ 2DGame/02-Bubble/02-Bubble/GUI.cpp
@@ -64,7 +64,7 @@ void GUI::update(int deltaTime)
     //Funcio que et digui els cors en funcio de la vida del personatge
 }
 
-void GUI::render(int vidatot)
+void GUI::render(float vidatot)
 {
     if (triar) {
         llança->render();
